Finished C judger.

+++ judge/model/submission.go
@@ -4,8 +4,8 @@ import "gorm.io/gorm"
 
 type Submission struct {
 	gorm.Model
-	ProblemID uint
-	UserID    uint
+	ProblemID uint32
+	UserID    uint16
 	Language  string
 	Code      string
 	Status    string
